stream-processing/kafka-consumer/requirements.txt
kafka-python
tabulate